feat(api): enforce route auth via decorator-driven middleware
Replace the per-route `requireMapping()` boilerplate with a single
`auth` Middleware piece (src/middlewares/auth.ts) that checks the
bearer token for any Route decorated with `@Authenticated()`
(src/lib/api/decorators.ts), attaching the resolved mapping as
`request.mappings`.
