+++ src/routes/experiments/[key]/index.get.ts
@@ -1,20 +1,18 @@
 import type { ApiRequest, ApiResponse } from "@wolfstar/plugin-api";
-import { requireMapping } from "#lib/api/auth";
+import { Authenticated } from "#lib/api/decorators";
 import { serializeExperiment } from "#lib/experiments";
 import { container } from "@sapphire/pieces";
 import { isNullish } from "@sapphire/utilities";
 import { HttpCodes } from "@wolfstar/http-framework";
 import { Route } from "@wolfstar/plugin-api";
 
+@Authenticated()
 export class ExperimentRoute extends Route {
 	public constructor(context: Route.LoaderContext) {
 		super(context, { name: "experiments-detail-get" });
 	}
 
 	public async run(request: ApiRequest, response: ApiResponse) {
-		const mappings = requireMapping(request, response);
-		if (mappings === null) return;
-
 		const key = request.params.key;
 
 		const experiment = await container.experiments.findById(key);
