+++ src/routes/experiments/[key]/index.patch.ts
@@ -1,6 +1,6 @@
 import type { Prisma } from "#generated/prisma";
 import type { ApiRequest, ApiResponse } from "@wolfstar/plugin-api";
-import { requireMapping } from "#lib/api/auth";
+import { Authenticated } from "#lib/api/decorators";
 import {
 	InvalidDate,
 	normalizeOptional,
@@ -14,15 +14,13 @@ import { isNullish } from "@sapphire/utilities";
 import { HttpCodes } from "@wolfstar/http-framework";
 import { Route } from "@wolfstar/plugin-api";
 
+@Authenticated()
 export class ExperimentUpdateRoute extends Route {
 	public constructor(context: Route.LoaderContext) {
 		super(context, { name: "experiments-update-patch" });
 	}
 
 	public async run(request: ApiRequest, response: ApiResponse) {
-		const mappings = requireMapping(request, response);
-		if (mappings === null) return;
-
 		const key = request.params.key;
 
 		let body: Record<string, unknown>;
