+++ src/routes/experiments/[key]/overrides/[entityType]/[entityId].delete.ts
@@ -1,19 +1,17 @@
 import type { ApiRequest, ApiResponse } from "@wolfstar/plugin-api";
-import { requireMapping } from "#lib/api/auth";
+import { Authenticated } from "#lib/api/decorators";
 import { toOverrideEntityType } from "#lib/experiments";
 import { container } from "@sapphire/pieces";
 import { HttpCodes } from "@wolfstar/http-framework";
 import { Route } from "@wolfstar/plugin-api";
 
+@Authenticated()
 export class ExperimentOverrideDeleteRoute extends Route {
 	public constructor(context: Route.LoaderContext) {
 		super(context, { name: "experiments-overrides-delete" });
 	}
 
 	public async run(request: ApiRequest, response: ApiResponse) {
-		const mappings = requireMapping(request, response);
-		if (mappings === null) return;
-
 		const { key, entityId } = request.params;
 
 		const entityType = toOverrideEntityType(request.params.entityType);
