+++ src/routes/experiments/[key]/resolve.get.ts
@@ -1,19 +1,17 @@
 import type { ApiRequest, ApiResponse } from "@wolfstar/plugin-api";
-import { requireMapping } from "#lib/api/auth";
+import { Authenticated } from "#lib/api/decorators";
 import { container } from "@sapphire/pieces";
 import { isNullishOrEmpty } from "@sapphire/utilities";
 import { HttpCodes } from "@wolfstar/http-framework";
 import { Route } from "@wolfstar/plugin-api";
 
+@Authenticated()
 export class ExperimentResolveRoute extends Route {
 	public constructor(context: Route.LoaderContext) {
 		super(context, { name: "experiments-resolve-get" });
 	}
 
 	public async run(request: ApiRequest, response: ApiResponse) {
-		const mappings = requireMapping(request, response);
-		if (mappings === null) return;
-
 		const key = request.params.key;
 
 		const entityType = request.query.get("entityType");
