+++ src/routes/guilds/[id].get.ts
@@ -1,18 +1,16 @@
 import type { ApiRequest, ApiResponse } from "@wolfstar/plugin-api";
-import { requireMapping } from "#lib/api/auth";
+import { Authenticated } from "#lib/api/decorators";
 import { container } from "@sapphire/pieces";
 import { HttpCodes } from "@wolfstar/http-framework";
 import { Route } from "@wolfstar/plugin-api";
 
+@Authenticated()
 export class GuildRoute extends Route {
 	public constructor(context: Route.LoaderContext) {
 		super(context, { name: "guilds-detail-get" });
 	}
 
 	public async run(request: ApiRequest, response: ApiResponse) {
-		const mappings = requireMapping(request, response);
-		if (mappings === null) return;
-
 		let id: bigint;
 		try {
 			id = BigInt(request.params.id);
@@ -26,8 +24,8 @@ export class GuildRoute extends Route {
 
 		const data = await container.prisma.guild.findFirst({
 			where: { id },
-			select: mappings.properties,
+			select: request.mappings.properties,
 		});
-		response.json(data ?? mappings.defaults, HttpCodes.OK);
+		response.json(data ?? request.mappings.defaults, HttpCodes.OK);
 	}
 }
